requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-tinymce==3.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.0
whitenoise==5.3.0
dnspython==2.1.0

appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.6
certifi==2020.12.5
charset-normalizer==2.0.7
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.4.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2020.11.15
pycodestyle==2.7.0
pylint==2.6.0
python-decouple==3.5
requests==2.26.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.7
virtualenv==20.4.0
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.4.0
